scripts/mask2former/exp04-GC/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# exp04 requirements  (exp03 からの新規追加ライブラリはなし)
# ─────────────────────────────────────────────────────────────────────────────

# ─────────────────────────────────────────────────────────────────────────────
# PyTorch (CUDA 12.4 ビルド)
# インストール時は pip install コマンドを分けて実行すること（下記セットアップ手順参照）
#   pip install torch==2.5.1 torchvision==0.20.1 --index-url https://download.pytorch.org/whl/cu124
# ─────────────────────────────────────────────────────────────────────────────
torch==2.5.1
torchvision==0.20.1

# ─────────────────────────────────────────────────────────────────────────────
# HuggingFace
# ─────────────────────────────────────────────────────────────────────────────
transformers==4.41.2
accelerate==0.30.0

# ─────────────────────────────────────────────────────────────────────────────
# 画像処理・データ拡張
# ─────────────────────────────────────────────────────────────────────────────
Pillow==10.3.0
albumentations>=2.0.0
albucore>=0.1.0
opencv-python-headless==4.9.0.80

# ─────────────────────────────────────────────────────────────────────────────
# 数値計算・可視化
# ─────────────────────────────────────────────────────────────────────────────
numpy==1.26.4
matplotlib==3.9.0

# ─────────────────────────────────────────────────────────────────────────────
# ユーティリティ
# ─────────────────────────────────────────────────────────────────────────────
tqdm==4.66.4
